requirements.txt
# Database
asyncpg>=0.29.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP Client (for API calls)
httpx>=0.25.0

# AI Integration
openai>=1.12.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Progress bars for CLI tools
tqdm>=4.65.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.6.0